정세인 - 테이블명세서 기준으로 수정

+++ zzonggeut/src/main/java/com/smhrd/myapp/domain/pet/PetRepository.java
@@ -3,6 +3,6 @@ package com.smhrd.myapp.domain.pet;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface PetRepository extends JpaRepository<PetEntity, Long> {
+public interface PetRepository extends JpaRepository<PetEntity, Integer> {
 	List findByUserId(Long userId);
 }
